WIP: Review Quotes refactored to use simplified scheme
