+++ client/src/Pages/Common/EditButtons.tsx
@@ -3,8 +3,8 @@ import jsonpatch from 'fast-json-patch';
 import useAxiosLater from 'Hooks/useAxiosLater';
 import useLogger from 'Hooks/useLogger';
 import useSelector from 'Hooks/useTypedSelector';
-import DeleteItemButton from 'Pages/Common/Modals/DeleteItemButton';
-import EditItemButton from 'Pages/Common/Modals/EditItemButtons';
+import DeleteItemButton from 'Pages/Common/Modals/_DeleteItemButton';
+import EditItemButton from 'Pages/Common/Modals/_EditItemButtons';
 import React, { ReactElement } from 'react';
 import { Item } from 'Types';
 import _ from 'underscore';
